today bootcamp

+++ contests/agc011/A/main.cpp
@@ -1,5 +1,5 @@
 #include <bits/stdc++.h>
-#define rep(i, n) for (int i = 0; i < n; i++)
+#define rep(i, n) for (int i = 0; i < (int)(n); i++)
 using namespace std;
 using ll = long long;
 using P = pair<int, int>;
